% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_FIPS_to_MSA}
\alias{add_FIPS_to_MSA}
\title{BRFSS}
\usage{
add_FIPS_to_MSA(data)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
