% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{add_population}
\alias{add_population}
\title{Add a population column to a GLP-format dataframe}
\usage{
add_population(df, geog)
}
\arguments{
\item{df}{A data frame}

\item{geog}{Title of the geography column}
}
\description{
Add a population column to a GLP-format dataframe
}
