% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wonder.R
\name{age_adj_rate}
\alias{age_adj_rate}
\title{Health Insurance}
\usage{
age_adj_rate(df, var, age_var = "age", pop_var = "population")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
