% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{any_time}
\alias{any_time}
\title{Read in a folder of CSV files}
\usage{
any_time(folder, starting_year = 2005, skip = 0, col_types = NULL)
}
\arguments{
\item{folder}{A path to a folder containing ACS data.}

\item{starting_year}{The first year for which there is data.}

\item{skip}{Nu8mber of lines to skip.}
}
\value{
A data frame
}
\description{
Read in a folder of CSV files
}
