% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{assign_col_join}
\alias{assign_col_join}
\title{Join two data frames or return one}
\usage{
assign_col_join(df_1, df_2, by)
}
\arguments{
\item{df_1}{A data frame that might exist}

\item{df_2}{A data frame to join to \code{df_1} column-wise}

\item{by}{Any values to pass to \code{full_join}.
If none are supplied, glptools::bind_df is used.}
}
\description{
Join two data frames or return one
}
