% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{assign_join}
\alias{assign_join}
\alias{assign_col_join}
\alias{assign_row_join}
\title{Join two data frames when one might not exist}
\usage{
assign_col_join(df_1, df_2, by)

assign_row_join(df_1, df_2)
}
\arguments{
\item{df_1}{A data frame that might exist}

\item{df_2}{A data frame to join to \code{df_1}}

\item{by}{If using \code{assign_col_join}, Any values to pass to \code{full_join}.
If none are supplied, \code{glptools::bind_df} is used.}
}
\description{
When accumulating data within a loop, you usually have to check whether to create a data frame
or bind two data frames together. This function is a shortcut: it checks to see is df_1 exists.
If so, it binds the two data frames together as usual. If not, it catches the error and returns df_2.
}
\section{Functions}{
\itemize{
\item \code{assign_col_join}: Joins two data frames column-wise when one might not exist

\item \code{assign_row_join}: Joins two data frames row-wise when one might not exist
}}

