% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{assign_row_join}
\alias{assign_row_join}
\title{Join two data frames or return one}
\usage{
assign_row_join(df_1, df_2)
}
\arguments{
\item{df_1}{A data frame that might exist}

\item{df_2}{A data frame to join to \code{df_1} row-wise}
}
\description{
Join two data frames or return one
}
