% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glp_utils.R
\name{bind_df}
\alias{bind_df}
\title{Joins data frames by common GLP ID variables}
\usage{
bind_df(..., by = NULL)
}
\arguments{
\item{...}{Data frames.}
}
\description{
Joins data frames by common GLP ID variables
}
