% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{block_group_to_council}
\alias{block_group_to_council}
\title{Convert census block group data to metro council district data}
\usage{
block_group_to_council(df, ...)
}
\arguments{
\item{df}{A data frame from the census}

\item{...}{a tidyselection of variables}
}
\description{
Convert census block group data to metro council district data
}
