% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{brfss_time}
\alias{brfss_time}
\title{Read in BRFSS data}
\usage{
brfss_time(folder)
}
\arguments{
\item{folder}{A path to a folder containing BRFSS data.}

\item{starting_year}{The first year for which there is data.}
}
\description{
Read in BRFSS data
}
