% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{business_time}
\alias{business_time}
\title{Read in County Business Pattern data}
\usage{
business_time(folder, starting_year = 2004)
}
\arguments{
\item{folder}{A path to a folder containing CBP data.}

\item{starting_year}{The first year for which there is data.}
}
\description{
Read in County Business Pattern data
}
