% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{census_to_council_direct}
\alias{census_to_council_direct}
\title{Convert census tract or block group data to metro council district data the simple way}
\usage{
census_to_council_direct(df, ..., geog)
}
\arguments{
\item{df}{A data frame from the census}

\item{...}{a tidyselection of variables}
}
\description{
Convert census tract or block group data to metro council district data the simple way
}
