% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microdata.R
\name{clean_cps_micro}
\alias{clean_cps_micro}
\title{Process CPS microdata from IPUMS.}
\usage{
clean_cps_micro(df, pull_peers = T)
}
\arguments{
\item{df}{A dataframe of microdata from IPUMS.}

\item{pull_peers}{Subset the data to peers? Defaults to \code{TRUE}.
Subsets to peer MSAs if present. Otherwise, subsets to peers counties.}
}
\description{
Process race, sex, and education variables if present.
}
