% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wonder.R
\name{clean_wonder}
\alias{clean_wonder}
\title{Health Insurance}
\usage{
clean_wonder(df)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
