% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs_helpers.R
\name{COLA}
\alias{COLA}
\title{COLA}
\usage{
COLA(df, ..., remove_calc = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
