% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cols_in}
\alias{cols_in}
\alias{\%cols_in\%}
\title{Returns any columns in data frame \code{df} that are in the character vector \code{columns},
  in the order the appear in \code{columns}}
\usage{
df \%cols_in\% columns
}
\description{
Returns any columns in data frame \code{df} that are in the character vector \code{columns},
  in the order the appear in \code{columns}
}
