% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{cols_not_in}
\alias{cols_not_in}
\alias{\%cols_not_in\%}
\title{Returns any columns in data frame \code{df} that are NOT in the character vector \code{columns},
  in the order the appear in \code{df}}
\usage{
df \%cols_not_in\% columns
}
\description{
Returns any columns in data frame \code{df} that are NOT in the character vector \code{columns},
  in the order the appear in \code{df}
}
