% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glp_utils.R
\name{complete_vector_arg}
\alias{complete_vector_arg}
\title{Expand a data frame to make missing combinations expicit}
\usage{
complete_vector_arg(df, columns, years, keep_empty_groups = FALSE)
}
\arguments{
\item{df}{description}

\item{columns}{description}

\item{years}{description}

\item{keep_empty_groups}{description}
}
\description{
Expand a data frame to make missing combinations expicit
}
