% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{df_type}
\alias{df_type}
\title{Organizes common GLP data by columns and rows.}
\usage{
df_type(df)
}
\arguments{
\item{df}{A data frame}
}
\description{
Organizes common GLP data by columns and rows.
}
