% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{get_sysdata}
\alias{get_sysdata}
\title{Add or replace a file in sysdata.rda
  Any files in the current environment are added to the sysdata.rda file}
\usage{
get_sysdata(df)
}
\description{
Add or replace a file in sysdata.rda
  Any files in the current environment are added to the sysdata.rda file
}
