% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%+\%}
\alias{\%+\%}
\title{BRFSS}
\usage{
a \%+\% b
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
