% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline.R
\name{graph_trendline_ky_ed}
\alias{graph_trendline_ky_ed}
\title{BRFSS}
\usage{
graph_trendline_ky_ed(df, var, plot_title = "", y_title = "Percent",
  caption_text = "", subtitle_text = "", rollmean = 1,
  break_settings = "", xmin = 2005, xmax = 2015)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
