% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline.R
\name{graph_trendline_lou}
\alias{graph_trendline_lou}
\title{BRFSS}
\usage{
graph_trendline_lou(df, var, plot_title = "", y_title = "Percent",
  caption_text = "", subtitle_text = "", rollmean = 1, xmin = 2005,
  xmax = 2016, break_settings = "")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
