% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline.R
\name{graph_trendline_race}
\alias{graph_trendline_race}
\title{BRFSS}
\usage{
graph_trendline_race(data_long, var = "var", value = "value",
  plot_title = "", y_title = "Percent", caption_text = "",
  subtitle_text = "", rollmean = 1, break_settings = "",
  xmin = 1996, xmax = 2016, labels, color_pal)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
