% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline.R
\name{graph_trendline_race_peer_two}
\alias{graph_trendline_race_peer_two}
\title{BRFSS}
\usage{
graph_trendline_race_peer_two(df, vars, plot_title = "",
  y_title = "Percent", peers = "Current", caption_text = "",
  subtitle_text = "", cat_labels = c("White", "Black"), rollmean = 1,
  xmin = 2005, xmax = 2016, break_settings = "")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
