% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance.R
\name{insurance_time}
\alias{insurance_time}
\title{Health Insurance}
\usage{
insurance_time(directory = "", starting.year = 2008)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
