% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ma_tract}
\alias{ma_tract}
\title{A crosswalk from tracts to market areas.}
\format{A data frame with two variables:
\describe{
\item{\code{GEO_ID}}{20-digit GEO_ID code}
\item{\code{market_area}}{market area name}
}}
\usage{
ma_tract
}
\description{
A crosswalk from tracts to market areas.
}
\keyword{datasets}
