% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline_helpers.R
\name{make_list}
\alias{make_list}
\title{Makes a named list of objects}
\usage{
make_list(...)
}
\arguments{
\item{...}{A list of objects}
}
\description{
The name of each object in the list matches its name in the environment.
}
