% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{make_map}
\alias{make_map}
\title{BRFSS}
\usage{
make_map(var, name, map_obj = "map_tract", geography = NULL,
  units = "Percent", map_style = "sequential", legend_title = "",
  palette = "", reverse = F, continuous = T, var_bins,
  tiles = TRUE, save_data = F, save_file = F, save_image = F,
  save_location)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
