% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{make_map_00}
\alias{make_map_00}
\title{BRFSS}
\usage{
make_map_00(var, name, units = "Percent", map_style = "sequential",
  legend_title = "")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
