% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_block_group}
\alias{map_block_group}
\title{A map of block groups in Louisville}
\format{
A SpatialPolygonsDataFrame
\describe{
\item{\code{GEO_ID}}{15-digit GEO_ID code consisting of FIPS, tract, and block numbers}
\item{\code{tract}}{6-character tract code}
\item{\code{block_group}}{numeric block group code}
\item{\code{line1}}{Block group \code{block}}
}
}
\usage{
map_block_group
}
\description{
A map of block groups in Louisville
}
\keyword{datasets}
