% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_county}
\alias{map_county}
\title{A map of counties in the Louisville MSA}
\format{
A SpatialPolygonsDataFrame
\describe{
\item{\code{FIPS}}{5-character FIPS code}
\item{\code{county}}{county name}
\item{\code{line1}}{\code{county}}
}
}
\usage{
map_county
}
\description{
A map of counties in the Louisville MSA
}
\keyword{datasets}
