% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_market}
\alias{map_market}
\title{A map of market areas in Louisville}
\format{A SpatialPolygonsDataFrame
\describe{
\item{\code{market}}{Market area}
\item{\code{line1}}{\code{Market area}}
}}
\usage{
map_market
}
\description{
A map of market areas in Louisville
}
\keyword{datasets}
