% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_tract_2000}
\alias{map_tract_2000}
\title{A map of 2000 tracts in Louisville}
\format{
A SpatialPolygonsDataFrame
\describe{
\item{\code{tract}}{6-character tract code}
\item{\code{line1}}{Tract \code{tract}}
}
}
\usage{
map_tract_2000
}
\description{
A map of 2000 tracts in Louisville
}
\keyword{datasets}
