% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_zip}
\alias{map_zip}
\title{A map of zip codes in Louisville}
\format{A SpatialPolygonsDataFrame
\describe{
\item{\code{zip}}{5-digit zip code}
\item{\code{line1}}{Zip code \code{zip}}
}}
\usage{
map_zip
}
\description{
A map of zip codes in Louisville
}
\keyword{datasets}
