% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSA_FIPS}
\alias{MSA_FIPS}
\title{A crosswalk of MSA codes to the FIPS codes of their core counties.}
\format{
A data frame with two variables:
\describe{
\item{\code{MSA}}{MSA code}
\item{\code{FIPS}}{FIPS code}
}
}
\usage{
MSA_FIPS
}
\description{
A crosswalk of MSA codes to the FIPS codes of their core counties.
}
\keyword{datasets}
