% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSA_names}
\alias{MSA_names}
\title{A crosswalk of MSA codes to MSA names.}
\format{A data frame with two variables:
\describe{
\item{\code{MSA}}{MSA code}
\item{\code{city}}{city}
}}
\usage{
MSA_names
}
\description{
A crosswalk of MSA codes to MSA names.
}
\keyword{datasets}
