% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSA_PUMA}
\alias{MSA_PUMA}
\title{A crosswalk of MSA codes to the PUMAs they contain.}
\format{A data frame with four variables:
\describe{
\item{\code{MSA}}{MSA code}
\item{\code{year}}{year.
  PUMAs based on the 2000 census are labelled with the years 2000 to 2012 to mirror ACS microdata available through IPUMS.
  PUMAS based on the 2010 census are labelled with the years 2013 to 2017.}
\item{\code{STATEFIP}}{State FIPS code}
\item{\code{PUMA}}{Public Use Microdata Area}
}}
\usage{
MSA_PUMA
}
\description{
A crosswalk of MSA codes to the PUMAs they contain.
}
\keyword{datasets}
