% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nh_tract}
\alias{nh_tract}
\title{A crosswalk from tracts to neighborhoods.}
\format{A data frame with two variables:
\describe{
\item{\code{GEO_ID}}{20-digit GEO_ID code}
\item{\code{neighborhood}}{neighborhood name}
}}
\usage{
nh_tract
}
\description{
A crosswalk from tracts to neighborhoods.
}
\keyword{datasets}
