% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{not_in}
\alias{not_in}
\alias{\%cols_in\%}
\alias{\%cols_not_in\%}
\title{Returns any variables in the vector
  that are columns in the data frame
  in the order the appear in the columns vector}
\usage{
df \%cols_in\% columns

df \%cols_not_in\% columns
}
\description{
Returns any variables in the vector
  that are columns in the data frame
  in the order the appear in the columns vector

Returns any variables in the vector
  that are NOT columns in the data frame.
}
