% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartype_utils.R
\name{pivot_vartype_longer}
\alias{pivot_vartype_longer}
\title{Convert census tract data to metro council district data}
\usage{
pivot_vartype_longer(
  df,
  names_from_col = "variable",
  values_from_col = "value"
)
}
\description{
Convert census tract data to metro council district data
}
