% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartype_utils.R
\name{pivot_vartype_wider}
\alias{pivot_vartype_wider}
\title{Convert census tract data to metro council district data}
\usage{
pivot_vartype_wider(df, ...)
}
\description{
Convert census tract data to metro council district data
}
