% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_df}
\alias{population_df}
\title{Population}
\format{A data frame with three variables:
\describe{
\item{\code{FIPS}}{FIPS codes}
\item{\code{year}}{year}
\item{\code{population}}{population}
}}
\usage{
population_df
}
\description{
Population data
}
\keyword{datasets}
