% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_df_merged}
\alias{population_df_merged}
\title{Population}
\format{A data frame with three variables:
\describe{
\item{\code{FIPS}}{FIPS codes (character)}
\item{\code{year}}{year}
\item{\code{population}}{population}
}}
\usage{
population_df_merged
}
\description{
Population data with merged and unmerged St. Louis counties and a character FIPS column.
}
\keyword{datasets}
