% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\name{process_acs}
\alias{process_acs}
\title{Process ACS microdata from IPUMS.}
\usage{
process_acs(df, gq = F, pull_peers = T, remove_vars = T)
}
\arguments{
\item{df}{A dataframe of microdata from IPUMS.}

\item{gq}{Include group quarters residents? Defaults to \code{FALSE}.}

\item{pull_peers}{Subset the data to peers? Defaults to \code{TRUE}.
Subsets to peer MSAs if present. Otherwise, subsets to peers counties.}
}
\description{
Adds FIPS codes and the Tulsa MSA code, if appropriate.
Process race, sex, and education variables if present.
}
