% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_api.R
\name{process_census}
\alias{process_census}
\title{Algorithm to create GLP-style data frame from census downloads}
\usage{
process_census(
  df,
  var_names = "value",
  cat_var,
  output_name,
  age_groups = "all",
  age_group_wide = FALSE,
  sum_to_100 = TRUE
)
}
\arguments{
\item{df}{A data frame from the census}

\item{var_names}{Variable names}

\item{cat_var}{Categorical vari}

\item{output_name}{Output var name}
}
\description{
Algorithm to create GLP-style data frame from census downloads
}
