% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\name{process_cps}
\alias{process_cps}
\title{Process CPS microdata from IPUMS.}
\usage{
process_cps(df, pull_peers = T)
}
\arguments{
\item{df}{A dataframe of microdata from IPUMS.}

\item{pull_peers}{Subset the data to peers? Defaults to \code{TRUE}.
Subsets to peer MSAs if present. Otherwise, subsets to peers counties.}
}
\description{
Process race, sex, and education variables if present.
}
