% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{process_map}
\alias{process_map}
\title{Create tract, nh, and muw map files from a map data frame}
\usage{
process_map(
  map_df,
  ...,
  return_name = NULL,
  maps = c("tract", "nh", "muw", "bg", "council"),
  council_MOE = TRUE
)
}
\arguments{
\item{map_df}{A data frame of census tracts with a 20-digit ID colum}

\item{...}{Variables to calculate}

\item{return_name}{The name of the maps to return. Suffixes will be appended.}

\item{maps}{The type of maps to return. Defaults to tract, nh, and muw.}

\item{council_MOE}{Simulate Metro Council Districts? Takes a while.}
}
\description{
Create tract, nh, and muw map files from a map data frame
}
