% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{process_map_og}
\alias{process_map_og}
\title{Create tract, nh, and muw map files from a map data frame}
\usage{
process_map_og(
  map_df,
  ...,
  pop,
  pop_adjust = F,
  return_name = NULL,
  method = "mean",
  maps = c("tract", "nh", "muw"),
  keep_pop = FALSE
)
}
\arguments{
\item{map_df}{A data frame of census tracts with a 20-digit ID colum}

\item{pop}{The population variable}

\item{return_name}{The name of the maps to return. Suffixes will be appended.}

\item{method}{The method used to aggregate tracts.
\code{percent} adds counts and populations to create a percentage
\code{mean} takes values and weights to create a weighted mean
\code{sum} adds counts and ignores the populations
\code{median} creates a weighted median}

\item{maps}{The type of maps to return. Defaults to tract, nh, and muw.}

\item{variables}{A character vector of variables}
}
\description{
Create tract, nh, and muw map files from a map data frame
}
