% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\name{process_microdata}
\alias{process_microdata}
\title{BRFSS}
\usage{
process_microdata(df, gq = F, pull_peers = T)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
