% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_lou.R
\name{pull_Lou}
\alias{pull_Lou}
\title{Subset a data frame containing MSA or county data to peer cities and add current and baseline peer data.}
\usage{
pull_Lou(df, geog = "", subset_to_Lou = T)
}
\arguments{
\item{df}{A data frame containing the column MSA or FIPS}

\item{geog}{Level of gegraphy to use for adding information and filtering.
Can be \code{FIPS} or \code{MSA}
Defaults to the MSA if present, followed by the FIPS column if present.
(To keep MSA-level data using FIPS as the ID column, geog should be \code{MSA}.)}

\item{subset_to_Lou}{Either \code{TRUE} or \code{FALSE}.
Subset the data to Louisville geographies? Defaults to \code{TRUE}.}
}
\description{
Subset a data frame containing MSA or county data to peer cities and add current and baseline peer data.
}
