% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pull_peers_MSA}
\alias{pull_peers_MSA}
\title{BRFSS}
\usage{
pull_peers_MSA(data)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
