% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{rank_and_nb_group}
\alias{rank_and_nb_group}
\title{Health Insurance}
\usage{
rank_and_nb_group(df, var, order = "Descending", peers = "Current",
  plot_title = "", y_title = "Percent", caption_text = "",
  sigfig = 3, num_dec = 1, text = TRUE, h_line = FALSE,
  subtitle_text = "", thousands_comma = T)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
