% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{rank_var}
\alias{rank_var}
\title{Health Insurance}
\usage{
rank_var(df_original, variables, order = "Descending",
  peers = "Current", new_vars = "")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
