% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{ranking_data}
\alias{ranking_data}
\title{Health Insurance}
\usage{
ranking_data(
  df,
  variables,
  years = "",
  sex = "total",
  race = "total",
  descending = TRUE,
  peers = "Current",
  new_vars = ""
)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
