% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{ranking_slope}
\alias{ranking_slope}
\title{Health Insurance}
\usage{
ranking_slope(dataframe, times, measurement, grouping, title = "",
  subtitle = "", caption = "", xtextsize = 36, ytextsize = 8,
  titletextsize = 48, subtitletextsize = 10, captiontextsize = 8,
  linethickness = 1, linecolor = "ByGroup", datatextsize = 10,
  datatextfamily = "sans", datatextface = "plain",
  labeltextfamily = "sans", labeltextface = "bold")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
