% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{readers}
\alias{readers}
\alias{any_time}
\alias{acs_time}
\alias{brfss_time}
\alias{business_time}
\alias{insurance_time}
\alias{unemployment_time}
\alias{wonder_time}
\title{GLP Readers}
\usage{
any_time(folder, starting_year = 2005, skip = 0, col_types = NULL, read_fxn)

acs_time(folder, geog = "FIPS", starting_year = 2005, additional_geogs = "")

brfss_time(folder)

business_time(folder, starting_year = 2004)

insurance_time(directory = "", starting.year = 2008)

unemployment_time(folder, starting_year = 2000)

wonder_time(folder, geog_type = "FIPS")
}
\arguments{
\item{folder}{A path to a folder containing files.
Years are assigned based on the file order, so the alphabetical order should match the chronological order.}

\item{starting_year}{The first year for which there is data.}

\item{skip}{Number of lines to skip from the top of the file}

\item{col_types}{Types of columns to pass to the reader functions}

\item{read_fxn}{Reader function}

\item{geog}{The geographic level of the data, usually \code{FIPS}, \code{MSA}, or \code{tract}}

\item{additional_geogs}{Additional geographies to be included in the final data frame.}
}
\value{
A data frame subset to peers
}
\description{
Read a folder of files to a data frame
}
\section{Functions}{
\itemize{
\item \code{any_time}: Reads in a folder of data

\item \code{acs_time}: Reads in a folder of ACS data

\item \code{brfss_time}: Reads in a folder of BRFSS data

\item \code{business_time}: Reads in a folder of County Business Pattern data

\item \code{insurance_time}: Reads in a folder of SAIHI data

\item \code{unemployment_time}: Reads in a folder of CPS unemployment data

\item \code{wonder_time}: Read in CDC Wonder data
Each file in the folder should be named according to the FIPS code of the data.
}}

\examples{
acs_time("bach_plus", additional_geogs = "21067")

}
