% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rollmean3}
\alias{rollmean3}
\title{Calculate the 3-year rolling mean of a vector
Deprecated for rollmeanr}
\usage{
rollmean3(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector of moving averages
}
\description{
Calculate the 3-year rolling mean of a vector
Deprecated for rollmeanr
}
