% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rollmeanr}
\alias{rollmeanr}
\title{Calculate the rolling mean of a vector}
\usage{
rollmeanr(x, r)
}
\arguments{
\item{x}{A vector}

\item{r}{The number of years to average}
}
\value{
A vector of rolling averages
}
\description{
Calculate the rolling mean of a vector
}
