% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wonder.R
\name{stl_adj_wonder}
\alias{stl_adj_wonder}
\title{Health Insurance}
\usage{
stl_adj_wonder(df, method = "sum", pop_var = "population")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
