% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stl_merge}
\alias{stl_merge}
\title{BRFSS}
\usage{
stl_merge(df_original, ..., weight_var = "", method = "mean")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
