% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartype_utils.R
\name{sum_by_var_type}
\alias{sum_by_var_type}
\title{Summarize vars}
\usage{
sum_by_var_type(df, ..., output_vartypes, drop_groups = TRUE)
}
\arguments{
\item{df}{.}

\item{...}{Column names to combine.}

\item{drop_groups}{.}
}
\description{
Summarize vars
}
