% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{sum_FIPS_to_MSA}
\alias{sum_FIPS_to_MSA}
\title{Subset a data frame containing MSA data to peer cities and add current and baseline peer data.}
\usage{
sum_FIPS_to_MSA(df)
}
\arguments{
\item{df}{A data frame containing the column MSA}
}
\description{
Subset a data frame containing MSA data to peer cities and add current and baseline peer data.
}
