% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microdata.R
\name{survey_by_demog}
\alias{survey_by_demog}
\title{Survey microdata by race and sex}
\usage{
survey_by_demog(
  df,
  var,
  weight_var = "PERWT",
  type = "categorical",
  geog = "FIPS",
  method = "default",
  other_grouping_vars = c(),
  breakdowns = c("total", "sex", "race", "sex_by_race")
)
}
\arguments{
\item{var}{A column name to perform svymean on.}

\item{weight_var}{weight variable}

\item{type}{Either mean or categorical}

\item{geog}{geographic column}

\item{other_grouping_vars}{other variables to group by}

\item{survey}{A survey object containing FIPS, year, and optional race and sex columns.}
}
\description{
Survey microdata by race and sex
}
