% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microdata.R
\name{svy_bootstrap}
\alias{svy_bootstrap}
\title{Survey microdata by a particular demographic grouping}
\usage{
svy_bootstrap(df, var, weight_var, type, grouping_vars)
}
\arguments{
\item{df}{A data frame of survey data}

\item{var}{The variable of interest}

\item{weight_var}{A column of weights}

\item{grouping_vars}{Columns to group by}
}
\description{
Survey microdata by a particular demographic grouping
}
