% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microdata.R
\name{svy_repwts}
\alias{svy_repwts}
\title{Survey using replicate weights}
\usage{
svy_repwts(df, var, weight_var, type, grouping_vars)
}
\arguments{
\item{df}{A data frame of survey data}

\item{var}{The variable of interest}

\item{weight_var}{A column of weights}

\item{grouping_vars}{Columns to group by}
}
\description{
Survey using replicate weights
}
