% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline_helpers.R
\name{tl_add_line_data}
\alias{tl_add_line_data}
\title{Add factor columns for ggplot to reference when assigning line groupings and styles}
\usage{
tl_add_line_data(df, type = "single", category = "", cat_names = "",
  drop_pctiles = F)
}
\value{
A data frame
}
\description{
Add factor columns for ggplot to reference when assigning line groupings and styles
}
