% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline_helpers.R
\name{tl_add_line_data_maxmin}
\alias{tl_add_line_data_maxmin}
\title{Add factor columns for ggplot to reference when assigning line groupings and styles for best and worst peer city graphs}
\usage{
tl_add_line_data_maxmin(df)
}
\value{
A data frame
}
\description{
Add factor columns for ggplot to reference when assigning line groupings and styles for best and worst peer city graphs
}
