% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline_helpers.R
\name{tl_break_settings}
\alias{tl_break_settings}
\title{Calculate x-axis break settings}
\usage{
tl_break_settings(df, xmin, xmax, rollmean)
}
\arguments{
\item{...}{df, xmin, xmax, rollmean}
}
\value{
major break settings and minor break settings
}
\description{
Major break settings describe where the x-axis labels fall.
Minor break settings describe where all other vertical lines fall.
}
