% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline_helpers.R
\name{tl_filter_cat}
\alias{tl_filter_cat}
\title{Filter data to category and create category names.}
\usage{
tl_filter_cat(df, type = "single", category = "",
  include_hispanic = F)
}
\value{
An object containing a data frame and category names
}
\description{
Filter data to category and create category names.
}
