% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline_helpers.R
\name{tl_rolling_mean}
\alias{tl_rolling_mean}
\title{Calculate the rolling mean of a data frame}
\usage{
tl_rolling_mean(df, xmin = 2000, xmax = 2017, rollmean = 1,
  census = T, subtitle_text = "")
}
\value{
A list containing df, xmin, xmax ,and subtitle_text
}
\description{
Calculate the rolling mean of a data frame
}
