% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline_helpers.R
\name{tl_style}
\alias{tl_style}
\title{Add style elements}
\usage{
tl_style(
  p,
  plot_title,
  y_title,
  caption_text,
  subtitle_text,
  cat_names,
  x_title = "Year"
)
}
\value{
A ggplot object
}
\description{
Add style elements
}
