% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline_helpers.R
\name{tl_year_breaks}
\alias{tl_year_breaks}
\title{Calculate the rolling mean of a data frame}
\usage{
tl_year_breaks(df, year_breaks)
}
\value{
A list containing df, xmin, xmax ,and subtitle_text
}
\description{
Calculate the rolling mean of a data frame
}
