% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{total_demographics}
\alias{total_demographics}
\title{Aggregates demographic data}
\usage{
total_demographics(df, ...)
}
\arguments{
\item{df}{A data frame}

\item{...}{Variables to total}
}
\description{
Aggregates demographic data
}
