% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{tract00_to10}
\alias{tract00_to10}
\title{Transform data from 2000 census tracts to 2010 census tracts}
\usage{
tract00_to10(df, years, ..., other_grouping_vars = "")
}
\description{
Transform data from 2000 census tracts to 2010 census tracts
}
