% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{tract10_to_00}
\alias{tract10_to_00}
\title{Transform data from 2010 census tracts to 2000 census tracts}
\usage{
tract10_to_00(df, years, ..., other_grouping_vars = "")
}
\description{
Transform data from 2010 census tracts to 2000 census tracts
}
