% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{tract_to_council}
\alias{tract_to_council}
\title{Convert census tract data to metro council district data}
\usage{
tract_to_council(df, ...)
}
\arguments{
\item{df}{A data frame from the census}

\item{...}{a tidyselection of variables}
}
\description{
Convert census tract data to metro council district data
}
