% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline.R
\name{trend}
\alias{trend}
\title{BRFSS}
\usage{
trend(df, var, cat = "", rollmean = 1, plot_title = "",
  y_title = "", caption_text = "", subtitle_text = "",
  peers = "current", ylimits = "", shading = F, pctiles = T,
  include_hispanic = F, include_asian = T, order = "descending",
  xmin = "", xmax = "")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
