% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline.R
\name{trend_ky_ed}
\alias{trend_ky_ed}
\title{BRFSS}
\usage{
trend_ky_ed(df, var, category = "", xmin = 2000, xmax = 2017,
  ylimits = "", rollmean = 1, plot_title = "", y_title = "",
  caption_text = "", subtitle_text = "", shading = F,
  drop_pctiles = F)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
