% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline.R
\name{trend_maxmin}
\alias{trend_maxmin}
\title{BRFSS}
\usage{
trend_maxmin(df, var, rollmean = 1, plot_title = "", y_title = "",
  caption_text = "", subtitle_text = "", peers = "current",
  ylimits = "", order = "descending", xmin = "", xmax = "")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
