% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline.R
\name{trend_maxmin_old}
\alias{trend_maxmin_old}
\title{BRFSS}
\usage{
trend_maxmin_old(df, var, sex = "total", race = "total",
  plot_title = "", y_title = "", peers = "Current",
  caption_text = "", subtitle_text = "", same_start = F,
  xmin = 2000, xmax = 2017, break_settings = "", rollmean = 1,
  output = "")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
