% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline.R
\name{trendline}
\alias{trendline}
\title{BRFSS}
\usage{
trendline(type, df, var, category, sex, race, peers, xmin, xmax, ylimits,
  rollmean, census, plot_title, y_title, caption_text, subtitle_text,
  shading, drop_pctiles, include_hispanic, order)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
