% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trendline_data}
\alias{trendline_data}
\title{BRFSS}
\usage{
trendline_data(df, rollmean = 1, exclude = "", census_2000 = T)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
