% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline.R
\name{trendline_multiple_old}
\alias{trendline_multiple_old}
\title{BRFSS}
\usage{
trendline_multiple_old(df, plot_title = "", vars,
  cat_labels = c("White", "Black", "Hispanic"), y_title = "Percent",
  peers = "Current", caption_text = "", subtitle_text = "",
  rollmean = 1, xmin = 2005, xmax = 2016, break_settings = "",
  shading = FALSE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
