% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline.R
\name{trendline_old}
\alias{trendline_old}
\title{BRFSS}
\usage{
trendline_old(df, var, sex = "total", race = "total",
  peers = "current", plot_title = "", y_title = "Percent",
  caption_text = "", subtitle_text = "", rollmean = 1, xmin = 2000,
  xmax = 2017, break_settings = "", census = T)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
