% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unemployment.R
\name{unemployment_time}
\alias{unemployment_time}
\title{BRFSS}
\usage{
unemployment_time(folder, starting_year = 2000)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
