% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{unfactor}
\alias{unfactor}
\title{Change a vector from a factor to character/numeric}
\usage{
unfactor(x)
}
\arguments{
\item{x}{A numeric or character vector encoded as a factor}
}
\value{
A numeric or character vector
}
\description{
If the vector can be converted to a numeric vector and fewer than 10% of the values become NA,
the vector is converted to a numeric vector. Otherwise it is converted to a character vector.
}
