% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{watterson_tract}
\alias{watterson_tract}
\title{Identifies which tracts are mostly inside the Watterson.}
\format{
A data frame with two variables:
\describe{
\item{\code{GEO_ID}}{20-digit GEO_ID code}
\item{\code{watterson}}{1 if the tract is mostly inside the Watterson, otherwise 0}
}
}
\usage{
watterson_tract
}
\description{
Identifies which tracts are mostly inside the Watterson.
}
\keyword{datasets}
