% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{weight_stl}
\alias{weight_stl}
\title{BRFSS}
\usage{
weight_stl(df_original, variables, weight_variable = "")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
