% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wonder.R
\name{wonder_time}
\alias{wonder_time}
\title{Health Insurance}
\usage{
wonder_time(folder, geog_type)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
