% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{wonder_time_age}
\alias{wonder_time_age}
\title{Read in CDC Wonder Data where each file is an age...deprecated?}
\usage{
wonder_time_age(folder, seq_var = "age", start = 0)
}
\arguments{
\item{folder}{A path to a folder containing CDC Wonder data.}

\item{seq_var}{The variable to sequence along.}

\item{start}{The first year of age in the data.}
}
\description{
Read in CDC Wonder Data where each file is an age...deprecated?
}
