% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wonder.R
\name{wonder_time_single}
\alias{wonder_time_single}
\title{Health Insurance}
\usage{
wonder_time_single(folder, seq_var = "age", start = 0)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
