% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glp_utils.R
\name{years_in_df}
\alias{years_in_df}
\title{Return the years in a data frame}
\usage{
years_in_df(df, var, category = "")
}
\arguments{
\item{df}{A data frame.}

\item{var}{A variable of interest.}

\item{demographic}{.
Defaults to \code{TRUE}.}
}
\description{
Return the years in a data frame
}
