% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{acs_time}
\alias{acs_time}
\title{Read in ACS data}
\usage{
acs_time(folder, geog = "FIPS", starting_year = 2005, additional_geogs = "")
}
\arguments{
\item{folder}{A path to a folder containing ACS data.}

\item{starting_year}{The first year for which there is data.}
}
\value{
A data frame
}
\description{
Read in ACS data
}
