% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\name{acs_time}
\alias{acs_time}
\title{BRFSS}
\usage{
acs_time(folder, starting_year = 2005)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
