% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{add_FIPS_to_MSA}
\alias{add_FIPS_to_MSA}
\title{Adds a FIPS column to a data frame containing a column of MSA codes.}
\usage{
add_FIPS_to_MSA(df)
}
\arguments{
\item{df}{A data frame containing the column MSA}
}
\description{
Adds a FIPS column to a data frame containing a column of MSA codes.
}
