% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wonder.R
\name{age_adj_rate}
\alias{age_adj_rate}
\title{Calculate an age-adjusted rate}
\usage{
age_adj_rate(df, var, age_var = "age", pop_var = "population")
}
\arguments{
\item{df}{A data frame.}

\item{var}{The variable to weight.}

\item{age_var}{Variable of age groups. Either "age" or "age_10" from \code{clean_wonder}.}

\item{pop_var}{The popylatuin variable.}
}
\description{
Calculate an age-adjusted rate
}
