% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{bind_df}
\alias{bind_df}
\title{Joins data frames by common GLP ID variables: FIPS, MSA, year, race, and sex.}
\usage{
bind_df(...)
}
\arguments{
\item{...}{Data frames.}
}
\description{
Joins data frames by common GLP ID variables: FIPS, MSA, year, race, and sex.
}
