% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bind_df}
\alias{bind_df}
\title{BRFSS}
\usage{
bind_df(...)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
