% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brfss.R
\name{brfss_time}
\alias{brfss_time}
\title{BRFSS}
\usage{
brfss_time(folder, start_year = 2002)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
