% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_api.R
\name{build_census_var_df}
\alias{build_census_var_df}
\title{Create a data frame to use for the Census API}
\usage{
build_census_var_df(survey, table, age_groups = F)
}
\arguments{
\item{survey}{The survey to get variables from.
Either \code{sf3}, \code{acs1}, or \code{acs5}.}

\item{table}{The table to return variables from.}

\item{age_groups}{Return variables broken down by age group?}
}
\description{
Create a data frame to use for the Census API
}
