% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbp.R
\name{business_time}
\alias{business_time}
\title{County Business Patterns}
\usage{
business_time(folder, starting_year = 2004)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
