% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{census_to_council}
\alias{census_to_council}
\title{Convert census block group data to metro council district data using simulations}
\usage{
census_to_council(df, ..., geog)
}
\arguments{
\item{df}{A data frame from the census}

\item{...}{a tidyselection of variables}
}
\description{
Convert census block group data to metro council district data using simulations
}
