% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wonder.R
\name{clean_wonder}
\alias{clean_wonder}
\title{Clean CDC Wonder data}
\usage{
clean_wonder(df, method = "sum")
}
\arguments{
\item{df}{A data frame.}

\item{method}{"sum" or "weight" to merge St. Louis.}
}
\description{
Processes a data frame of CDC Wonder mortality or natality data.
Renames and selects colums of interest, recodes age groups,
filters out rows with totals, recodes race variables, and merges St. Louis counties
}
