% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_utils.R
\name{COLA}
\alias{COLA}
\title{Adjust data for cost of living and inflation}
\usage{
COLA(df, ..., base_year = 2019, remove_calc = TRUE, inflation = T, rpp = T)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Variables to adjust.}

\item{remove_calc}{Whether to remove the columns \code{rpp_index} and \code{cpi_index} after adjustments.
Defaults to \code{TRUE}.}
}
\description{
Adjust data for cost of living and inflation
}
