% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glp_utils.R
\name{complete_vector_arg}
\alias{complete_vector_arg}
\title{Load GLP-related packages}
\usage{
complete_vector_arg(df, columns, years, keep_empty_groups = FALSE)
}
\arguments{
\item{graphs}{Will graphs or maps be made?}
}
\description{
Load GLP-related packages
}
