% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_api.R
\name{get_census}
\alias{get_census}
\title{Get data from the Census API}
\usage{
get_census(var_df, geog, parallel = F)
}
\arguments{
\item{var_df}{A data frame of variables to send to the census API
produced by build_census_var_df}

\item{geog}{The geography to return variables for}

\item{parallel}{Use future and furrr?}
}
\description{
Get data from the Census API
}
