% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glp_utils.R
\name{glp_load_packages}
\alias{glp_load_packages}
\title{Load GLP-related packages}
\usage{
glp_load_packages(graphs = F)
}
\arguments{
\item{graphs}{Will graphs or maps be made?}
}
\description{
Load GLP-related packages
}
