% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{insurance_time}
\alias{insurance_time}
\title{Read in health insurance data}
\usage{
insurance_time(directory = "", starting.year = 2008)
}
\arguments{
\item{folder}{A path to a folder containing SAIHI data.}

\item{starting_year}{The first year for which there is data.}
}
\description{
Read in health insurance data
}
