% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{make_map}
\alias{make_map}
\title{BRFSS}
\usage{
make_map(var, name, map, legend_title, units = "Percent",
  map_style = "sequential", palette = "", reverse = F,
  continuous = T, var_bins, tiles = T, save_file = "",
  save_image = "")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()
}
\keyword{cats}
