% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{organize}
\alias{organize}
\title{Organizes common GLP data by columns and rows and replaces FIPS 1073 with 01073.}
\usage{
organize(df)
}
\arguments{
\item{df}{A data frame}
}
\description{
Columns: MSA, FIPS, city, year, sex, race, baseline, current,
Rows: MSA, FIPS, year, sex, race
}
