% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{per_capita_adj}
\alias{per_capita_adj}
\title{Create population-adjusted variables}
\usage{
per_capita_adj(
  df,
  ...,
  pop_var = "population",
  geog,
  keep_vars = T,
  keep_pop = F
)
}
\arguments{
\item{df}{A data frame}

\item{...}{Variables}

\item{geog}{The geography}

\item{keep_vars}{Keep original variables?}

\item{keep_pop}{Keep population in data frame}
}
\description{
Create population-adjusted variables
}
