% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{process_map}
\alias{process_map}
\title{Create tract, nh, and muw map files from a map data frame}
\usage{
process_map(map_df, variables, pop, method = "percent",
  return_name = NULL, maps = c("tract", "nh", "muw"))
}
\description{
Create tract, nh, and muw map files from a map data frame
}
