% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{pull_peers_FIPS}
\alias{pull_peers_FIPS}
\title{Subset a data frame containing MSA data to peer cities and add current and baseline peer data.}
\usage{
pull_peers_FIPS(df, add_info = TRUE)
}
\arguments{
\item{df}{A data frame containing the column MSA}

\item{add_info}{Add city names, current peer, and baseline peer columns? Defaults to TRUE.}
}
\description{
Subset a data frame containing MSA data to peer cities and add current and baseline peer data.
}
