% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{pull_peers_MSA}
\alias{pull_peers_MSA}
\title{Subset a data frame containing MSA data to peer cities and add current and baseline peer data.}
\usage{
pull_peers_MSA(df, add_info = T)
}
\arguments{
\item{df}{A data frame containing the column MSA}
}
\description{
Subset a data frame containing MSA data to peer cities and add current and baseline peer data.
}
