% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{reshape_sex}
\alias{reshape_sex}
\title{Gathers data from wide format into long format based on sex.}
\usage{
reshape_sex(df)
}
\arguments{
\item{df}{A data frame}
}
\description{
Variables should end in ".male", ".female", or "" (blank for totals).
Variable names should not contain other periods.
}
