% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rollmean3}
\alias{rollmean3}
\title{BRFSS}
\usage{
rollmean3(x)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
