% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{rollmean5}
\alias{rollmean5}
\title{Calculate the 5-year rolling mean of a vector
Deprecated for rollmeanr}
\usage{
rollmean5(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector of moving averages
}
\description{
Calculate the 5-year rolling mean of a vector
Deprecated for rollmeanr
}
