% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wonder.R
\name{stl_adj_wonder}
\alias{stl_adj_wonder}
\title{Merge St. Louis}
\usage{
stl_adj_wonder(df, method = "sum", pop_var = "population")
}
\arguments{
\item{df}{A data frame.}

\item{method}{"sum" or "weight" to merge St. .}

\item{pop_var}{The population variable.}
}
\description{
Merge St. Louis
}
