% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_utils.R
\name{stl_merge}
\alias{stl_merge}
\title{Combines rows of data from the two St. Louis counties into one.}
\usage{
stl_merge(
  df,
  ...,
  weight_var = "",
  method = "",
  other_grouping_vars = "",
  just_replace_FIPS = F,
  keep_counties = F
)
}
\arguments{
\item{...}{Column names to combine.}

\item{weight_var}{A variable to use when weighting the counties. Defaults to \code{population}.}

\item{method}{"mean" or "sum". Defaults to mean.}

\item{df_original}{A data frame.}
}
\description{
Combines rows of data from the two St. Louis counties into one.
}
