% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{stl_merge}
\alias{stl_merge}
\title{Combines rows of data from the two St. Louis counties into one.}
\usage{
stl_merge(df_original, ..., weight_var = "", method = "mean")
}
\arguments{
\item{df_original}{A data frame.}

\item{...}{Column names to combine.}

\item{weight_var}{A variable to use when weighting the counties. Defaults to \code{population}.}

\item{method}{"mean" or "sum". Defaults to mean.}
}
\description{
Combines rows of data from the two St. Louis counties into one.
}
