% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{sum_FIPS_to_MSA}
\alias{sum_FIPS_to_MSA}
\title{Group data by MSA and summarise}
\usage{
sum_FIPS_to_MSA(df, ..., other_grouping_vars = "")
}
\arguments{
\item{df}{A data frame containing the column MSA}
}
\description{
Group data by MSA and summarise
}
