% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\name{svy_race_sex}
\alias{svy_race_sex}
\title{Survey microdata by race and sex}
\usage{
svy_race_sex(survey, var, race = T, sex = T, cross = T)
}
\arguments{
\item{survey}{A survey object containing FIPS, year, and optional race and sex columns.}

\item{var}{A column name to perform svymean on.}

\item{race}{Break down by race? Defaults to \code{TRUE}.}

\item{sex}{Break down by sex? Defaults to \code{TRUE}.}

\item{cross}{Break down by race and sex? Defaults to \code{TRUE}.}
}
\description{
This function uses the survey package to break down estimates by demographic group.
}
