% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\name{svy_race_sex}
\alias{svy_race_sex}
\title{Survey by race and sex}
\usage{
svy_race_sex(survey, var, race = T, sex = T, cross = T)
}
\arguments{
\item{survey}{A survey object containing FIPS, year, and optional race and sex columns.}

\item{var}{A column name to perform svymean on.}

\item{race}{Break down by race?}

\item{sex}{Break down by sex?}

\item{cross}{Break down by race and sex?}
}
\description{
Survey by race and sex
}
