% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline_helpers.R
\name{tl_filter}
\alias{tl_filter}
\title{Filter data to sex, race, and peer set.}
\usage{
tl_filter(df, var, peers, cat, include_hispanic, include_asian,
  return_only_df = F)
}
\value{
A data frame
}
\description{
Filter data to sex, race, and peer set.
}
