% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline_helpers.R
\name{tl_reshape_data}
\alias{tl_reshape_data}
\title{Reshape data to long format}
\usage{
tl_reshape_data(df)
}
\value{
A data frame containing the columns year, --category--, variable, and value.
}
\description{
Reshape data to long format
}
