% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline_helpers.R
\name{tl_rolling_mean}
\alias{tl_rolling_mean}
\title{Calculate the rolling mean of a data frame}
\usage{
tl_rolling_mean(df, xmin, xmax, rollmean, subtitle_text, return_only_df = F)
}
\value{
A list containing df, xmin, xmax ,and subtitle_text
}
\description{
Calculate the rolling mean of a data frame
}
