% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glp_utils.R
\name{total_demographics}
\alias{total_demographics}
\title{Aggregates demographic data}
\usage{
total_demographics(
  df,
  ...,
  total_sex = T,
  total_race = F,
  include_na = F,
  other_grouping_vars = "",
  moe = T
)
}
\arguments{
\item{df}{A data frame}

\item{...}{Variables to total}
}
\description{
Aggregates demographic data
}
