% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{tract_00_to_10}
\alias{tract_00_to_10}
\title{Transform data from 2000 census tracts to 2010 census tracts}
\usage{
tract_00_to_10(df, years, ..., other_grouping_vars = "")
}
\description{
Transform data from 2000 census tracts to 2010 census tracts
}
