% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{tract_to_council}
\alias{tract_to_council}
\title{Convert census tract data to metro council district data}
\usage{
tract_to_council(df, ..., method = "percent", weight_var = "total")
}
\arguments{
\item{df}{A data frame from the census}

\item{var_names}{Variable names}

\item{cat_var}{Categorical vari}

\item{output_name}{Output var name}
}
\description{
Convert census tract data to metro council district data
}
