% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline.R
\name{trend_data}
\alias{trend_data}
\title{Returns statistics calculated for use in trendline graphs}
\usage{
trend_data(df, var = "var", rollmean = 1, cat = "",
  peers = "current")
}
\arguments{
\item{df}{A data frame}

\item{rollmean}{A rolling mean. Defaults to 1.}

\item{census_2000}{Is the 2000 data from the census? If so, do not use rollmeanr on data from 2000. Defaults to TRUE.}
}
\description{
Returns statistics calculated for use in trendline graphs
}
