% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendline.R
\name{trend_maxmin}
\alias{trend_maxmin}
\title{BRFSS}
\usage{
trend_maxmin(df, var, order = "descending", sex = "total",
  race = "total", peers = "current", xmin = 2000, xmax = 2017,
  ylimits = "", rollmean = 1, census = T, plot_title = "",
  y_title = "", caption_text = "", subtitle_text = "")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Reads in BRFSS data
}
\examples{
cat_function()

}
\keyword{cats}
