% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trendline_data}
\alias{trendline_data}
\title{Returns statistics calculated for use in trendline graphs}
\usage{
trendline_data(df, rollmean = 1, race_type = "total",
  sex_type = "total", census_2000 = T)
}
\arguments{
\item{df}{A data frame}

\item{rollmean}{A rolling mean. Defaults to 1.}

\item{census_2000}{Is the 2000 data from the census? If so, do not use rollmeanr on data from 2000. Defaults to TRUE.}
}
\description{
Returns statistics calculated for use in trendline graphs
}
