% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{unemployment_time}
\alias{unemployment_time}
\title{Read in unemployment data}
\usage{
unemployment_time(folder, starting_year = 2000)
}
\arguments{
\item{folder}{A path to a folder containing CPS unemployment data.}

\item{starting_year}{The first year for which there is data.}
}
\description{
Read in unemployment data
}
