% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{weight_stl}
\alias{weight_stl}
\title{Combines rows of data from the two St. Louis counties into one. Requires variables to be in a character vector.
Deprecated in favor of stl_merge.}
\usage{
weight_stl(df_original, variables, weight_variable = "")
}
\arguments{
\item{df_original}{A data frame}

\item{variables}{A character vector of column names to combine}

\item{weight_var}{A variable to use when weighting the counties. If none is provided, population data from the ACS is used.}
}
\description{
Combines rows of data from the two St. Louis counties into one. Requires variables to be in a character vector.
Deprecated in favor of stl_merge.
}
