% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wonder.R
\name{wonder_time}
\alias{wonder_time}
\title{Read in CDC Wonder data}
\usage{
wonder_time(folder, geog_type = "FIPS")
}
\arguments{
\item{folder}{A path to a folder containing CDC Wonder data.}

\item{starting_year}{The first year for which there is data.}
}
\description{
Each file in the folder should be named according to the FIPS code of the data.
}
