% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wonder.R
\name{wonder_time_single}
\alias{wonder_time_single}
\title{Read in CDC Wonder Data}
\usage{
wonder_time_single(folder, seq_var = "age", start = 0)
}
\arguments{
\item{folder}{A path to a folder containing CDC Wonder data.}

\item{seq_var}{The variable to sequence along.}

\item{start}{The first year of age in the data.}
}
\description{
Each file in the folder should be named according to the age included in the data.
}
